+++ packages/algolia-tools-cli-generator/template/tasks/firebase/index.ts
@@ -1,13 +1,13 @@
 import algoliaModule from '../../account/algoliaAccount'
 import { useFirebaseAccount } from '../../account/firebaseAccount'
-import FirebaseInvoke from '../../../src/plugin/FirebaseInvoke'
+import FirestorePlugin from '../../../src/plugin/FirestorePlugin'
 import { createFirestoreCommanderPlugin } from '../../../src/plugin/generate-commands/firestore'
 import { Command } from 'commander'
 
 const algoliaIndexManager = algoliaModule.algoliaIndexManager
 
 const firebaseManager = useFirebaseAccount
-  ? new FirebaseInvoke({
+  ? new FirestorePlugin({
       algoliaIndexManager,
     })
   : undefined