+++ packages/algolia-tools/src/types.ts
@@ -1,4 +1,5 @@
 import { AlgoliaIndexManager } from './util/AlgoliaIndexManager'
+
 export interface IndexInterface<T extends any = any> {
   deleteIndexData(id: string | string[]): Promise<boolean>
   sendIndex(id: string, data: T): Promise<boolean>
@@ -13,12 +14,12 @@ export interface IndexConstructor<Ex = unknown> {
   ): IndexInterface
 }
 
-export interface AlgoliaToolsModule {
+export type AlgoliaToolsModule<Ex = unknown> = {
   algoliaIndexManager: AlgoliaIndexManager
   indices: {
     [collectionName: string]: IndexInterface
   }
-}
+} & Ex
 
 export type ExPlugin<PluginId extends string, PluginType> = {
   id: PluginId