+++ CHANGELOG.md
@@ -213,7 +213,7 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Features
 
-* getIndexName interface and export FirebaseInvoke ([fd26ceb](https://github.com/MOCHI-inc-JAPAN/algolia-firebase-tools/commit/fd26cebeefbb59fb2efcee531ed51b91d8f6cb5d))
+* getIndexName interface and export FirestorePlugin ([fd26ceb](https://github.com/MOCHI-inc-JAPAN/algolia-firebase-tools/commit/fd26cebeefbb59fb2efcee531ed51b91d8f6cb5d))
 
 ### [1.8.1](https://github.com/MOCHI-inc-JAPAN/algolia-firebase-tools/compare/v1.8.0...v1.8.1) (2020-08-19)